requirements.txt
altair==4.2.0
attrs==22.1.0
autopep8==1.7.0
azure-ai-textanalytics==5.2.0b2
azure-cognitiveservices-speech==1.23.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
fonttools==4.38.0
fpdf2==2.5.7
gitdb==4.0.9
GitPython==3.1.29
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
moviepy==1.0.3
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
proglog==0.1.10
protobuf==3.20.3
pyarrow==10.0.0
pycodestyle==2.9.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.14.0
svg.path==6.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
zipp==3.10.0
